chore: relocate files proper
