+++ app/src/main/java/campus/tech/kakao/map/model/SearchKeyword.kt
@@ -1,4 +1,4 @@
-package campus.tech.kakao.map
+package campus.tech.kakao.map.model
 
 import android.provider.BaseColumns
