+++ app/src/main/java/campus/tech/kakao/map/model/SearchResults.kt
@@ -1,4 +1,4 @@
-package campus.tech.kakao.map
+package campus.tech.kakao.map.model
 
 import com.google.gson.annotations.SerializedName
