+++ app/src/main/java/campus/tech/kakao/map/repository/KakaoSearchKeywordAPI.kt
@@ -1,5 +1,6 @@
-package campus.tech.kakao.map
+package campus.tech.kakao.map.repository
 
+import campus.tech.kakao.map.model.SearchResults
 import retrofit2.http.GET
 import retrofit2.http.Header
 import retrofit2.http.Query
