+++ app/src/main/java/campus/tech/kakao/map/view/SavedSearchKeywordsAdapter.kt
@@ -1,10 +1,11 @@
-package campus.tech.kakao.map
+package campus.tech.kakao.map.view
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
 import campus.tech.kakao.map.databinding.SavedSearchKeywordItemBinding
+import campus.tech.kakao.map.model.SearchKeyword
 
 class SavedSearchKeywordsAdapter(
     private val savedSearchKeywords: List<SearchKeyword>,
