+++ app/src/main/res/layout/activity_search_window.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:padding="10dp"
-    tools:context=".SearchWindowActivity">
+    tools:context=".view.SearchWindowActivity">
 
     <EditText
         android:id="@+id/searchWindow"
